Changes to be able to build multiple firmwares
Change to use the C preprocessor (use the ARM one for that) for
asm files so we can force-include a per-machine GPIO definition
file and use some preprocessor macros to define the cache registers.

This allows us to deal with the cases where the clock and data
use a different cache vs. the same cache.

Signed-off-by: Benjamin Herrenschmidt <benh@kernel.crashing.org>