Remove TODO comments in clock_out_bit and document optimisation
The send path isn't the most critical and optimizing here
is hard for little return. Also on systems like Romulus where
both clk and data are in the same GPIO register, we already
don't need the test for value change since we use a single
write to write both value and clock.

Signed-off-by: Benjamin Herrenschmidt <benh@kernel.crashing.org>