Fix problems with GPIO usage
The "value" register doesn't return the last data written when
read but returns the value sampled on the line, which may or
may not be the value written for an output GPIO. For example
it can be delayed by the synchronizers.

So whenever the CF reads the value to adjust its caches it needs
to use the "data read" register intead, otherwise it can clobber
values set by Linux when doing handshaking.

Also don't use %d0 for DTRA, use %d1 to avoid interrupts clobber
and do read/modify/write cycles on it as it's not a permanent
cached value, just a temporary.

Signed-off-by: Benjamin Herrenschmidt <benh@kernel.crashing.org>