Rework interfaces
We now use a portion of the ucode itself for version information
and to configure the base of SRAM, this will allow the driver
to use a pool allocator with less constraints.

We put the status register as part of the command register, this
speeds things up a bit and remove the need to clear it from
the host.

We also turn some SRAM stores from byte to long stores, they seem
to be faster (still room for improvement)

Signed-off-by: Benjamin Herrenschmidt <benh@kernel.crashing.org>